Name refactoring and some more code tidy up
