+++ src/testing/CMakeLists.txt
@@ -1,9 +1,9 @@
-add_subdirectory(Thymio_diffusion)
+add_subdirectory(thymio_diffusion)
 
 if(ARGOS_BUILD_FOR STREQUAL "thymio")
     add_subdirectory(sensor_test)
     add_subdirectory(sensor_test2)
     add_subdirectory(ground_sensor_test)
-    add_subdirectory(Accelerometer_test)
-    add_subdirectory(Accelerometer_test2)
+    add_subdirectory(accelerometer_test)
+    add_subdirectory(accelerometer_test2)
 endif(ARGOS_BUILD_FOR STREQUAL "thymio")
