+++ src/testing/accelerometer_test/accelerometer_test.cpp
@@ -1,5 +1,5 @@
 /* Include the controller definition */
-#include "accelerometer_test2.h"
+#include "accelerometer_test.h"
 /* Function definitions for XML parsing */
 #include <argos3/core/utility/configuration/argos_configuration.h>
 /* 2D vector definition */
@@ -21,7 +21,7 @@ CAccelerometerTest::CAccelerometerTest() :
    m_pcLeds(NULL),
    m_cAlpha(10.0f),
    m_fDelta(0.5f),
-   m_fWheelVelocity(20),
+   m_fWheelVelocity(0),
    m_cGoStraightAngleRange(-ToRadians(m_cAlpha),
                            ToRadians(m_cAlpha)) {}
 
@@ -32,51 +32,84 @@ void CAccelerometerTest::Init(TConfigurationNode& t_node) {
 
    m_pcWheels           = GetActuator<CCI_DifferentialSteeringActuator>("differential_steering");
    m_pcLeds             = GetActuator<CCI_ThymioLedsActuator          >("thymio_led");
-   m_pcAccelerometer    = GetSensor  <CCI_Thymio_acc_sensor           >("Thymio_accelerometer");
+   m_pcAccelerometer    = GetSensor  <CCI_Thymio_acc_sensor           >("thymio_accelerometer");
 
    std::cout<<"accelerometer: "<<m_pcAccelerometer;
 
    start = std::clock();
+   speedincrease = 5;
    timer = 0;
    rate = 20;
-   second = 0.00f;
-
+   second = 1.00f;
    try {
-       acc_sensor_readings.open("acc_sensor_readings.csv");
-   } catch (std::exception e) {
+       acc_sensor_readings.open("accelerometer_test_run.csv");
+   } catch (std::exception& e) {
        std::cout << e.what();
    }
+
+   m_fWheelVelocity = 0;
    m_pcLeds->SetProxHIntensity({32,32,32,32,32,32,32,32});
    m_pcLeds->SetProxHIntensity({0,0,0,0,0,0,0,0});
-
 }
 
 /****************************************/
 /****************************************/
 
 void CAccelerometerTest::ControlStep() {
-
     std::cout<< m_pcAccelerometer->accValues.x<<" ";
     std::cout<< m_pcAccelerometer->accValues.y<<" ";
     std::cout<< m_pcAccelerometer->accValues.z<<"\n";
 
+    m_pcLeds->SetProxHIntensity({32,32,32,32,32,32,32,32});
+    m_pcLeds->SetProxHIntensity({0,0,0,0,0,0,0,0});
+
+
     if(timer > second )
     {
-        if(second < 10)
+        if(second <5 )
         {
-            m_pcWheels->SetLinearVelocity(0,0);
+            std::cout<<"wait\t";
         }
-        else if(second < 30){
-            m_pcWheels->SetLinearVelocity(m_fWheelVelocity,-m_fWheelVelocity);//clockwise
+        else if(second < 10)
+        {
+            std::cout<<"Speed Up!!!\t";
+            m_fWheelVelocity += speedincrease;
+            m_pcWheels->SetLinearVelocity(m_fWheelVelocity,m_fWheelVelocity);
         }
-        else if(second < 40){
-            m_pcWheels->SetLinearVelocity(0,0);
+        else if(second < 20 )
+        {
+            std::cout<<"wait\t";
         }
-        else if(second < 60){
-            m_pcWheels->SetLinearVelocity(-m_fWheelVelocity,m_fWheelVelocity);//counter clockwise
+        else if(second < 25 )
+        {
+            std::cout<<"slow down!!!\t";
+            m_fWheelVelocity -= speedincrease;
+            m_pcWheels->SetLinearVelocity(m_fWheelVelocity,m_fWheelVelocity);
         }
-        else if(second < 70){
-            m_pcWheels->SetLinearVelocity(0,0);
+        else if(second < 35 )
+        {
+            std::cout<<"wait\t";
+        }
+        else if(second < 40)
+        {
+            std::cout<<"slow down!!!\t";
+            m_fWheelVelocity -= speedincrease;
+            m_pcWheels->SetLinearVelocity(m_fWheelVelocity,m_fWheelVelocity);
+
+        }
+        else if(second < 50)
+        {
+            std::cout<<"wait\t";
+        }
+        else if(second < 55)
+        {
+            std::cout<<"Speed Up!!!\t";
+            m_fWheelVelocity += speedincrease;
+            m_pcWheels->SetLinearVelocity(m_fWheelVelocity,m_fWheelVelocity);
+        }
+        else if(second < 65)
+        {
+            std::cout<<"wait\t";
         }
         else
         {
@@ -85,7 +118,7 @@ void CAccelerometerTest::ControlStep() {
                 acc_sensor_readings<<"\n";
                 acc_sensor_readings.close();
                 std::cout<< "File is exported! end of the experiment!";
-            } catch (std::exception e) {
+            } catch (std::exception& e) {
                 std::cout << e.what();
             }
             std::cin>> c;
@@ -93,8 +126,8 @@ void CAccelerometerTest::ControlStep() {
 
         second += 1.00f; // means a second has passed add to second to prepare it for next second
         std::cout<<second<<"\t One second passed!!!\n";
-    }
 
+    }
     acc_sensor_readings << timer<<",";
     acc_sensor_readings << m_pcAccelerometer->accValues.x<<",";
     acc_sensor_readings << m_pcAccelerometer->accValues.y<<",";
@@ -103,6 +136,9 @@ void CAccelerometerTest::ControlStep() {
     timer = (( std::clock() - start )*100/(double) CLOCKS_PER_SEC);
 //    std::cout<<( std::clock() - start )/(double) CLOCKS_PER_SEC;
     std::cout<<timer<<") ";
+    std::cout<<m_fWheelVelocity<<"\t";
+
+
 
 }
 
@@ -114,4 +150,4 @@ CAccelerometerTest::~CAccelerometerTest(){
 /****************************************/
 /****************************************/
 
-REGISTER_CONTROLLER(CAccelerometerTest, "accelerometer_test2_controller")
+REGISTER_CONTROLLER(CAccelerometerTest, "accelerometer_test_controller")
