+++ src/testing/accelerometer_test/color.png
