+++ src/testing/accelerometer_test2/color.png
