+++ src/testing/sensor_test2/testexperiment.argos
@@ -24,8 +24,8 @@
       </actuators>
 
       <sensors>
-        <Thymio_proximity implementation="default" show_rays="true" />
-	<Thymio_ground implementation="rot_z_only" show_rays="false"/>
+        <thymio_proximity implementation="default" show_rays="true" />
+	<thymio_ground implementation="rot_z_only" show_rays="false"/>
       </sensors>
 
       <params alpha="0" delta="0.4" velocity="1" />
