+++ src/testing/thymio_diffusion/color.jpg
