+++ src/testing/thymio_diffusion/color.png
