+++ src/testing/thymio_diffusion/loop_functions/draw_position.h
