+++ src/testing/thymio_diffusion/trajectory_loop_functions/trajectory_loop_functions.h
