+++ src/testing/thymio_diffusion/trajectory_loop_functions/trajectory_qtuser_functions.cpp
